+++ .eslintrc.json
@@ -15,9 +15,7 @@
     "@typescript-eslint/no-var-requires": "off",
     "react/function-component-definition": "off",
     "arrow-body-style": "off",
-    "react/button-has-type": "off",
     "react/no-unknown-property": ["error", { "ignore": ["css"] }],
-    "@typescript-eslint/no-unused-vars": "off",
     "@typescript-eslint/no-use-before-define": "off",
     "react/react-in-jsx-scope": "off",
     "import/no-absolute-path": "off"
